Move similar ideas to different folder
