+++ front/app/containers/IdeasNewPage/SimilarIdeas/messages.ts
@@ -2,20 +2,20 @@ import { defineMessages } from 'react-intl';
 
 export default defineMessages({
   similarSubmissionsPosted: {
-    id: 'app.containers.ideas.similarIdeas.similarSubmissionsPosted',
+    id: 'app.components.ideas.similarIdeas.similarSubmissionsPosted',
     defaultMessage: 'Similar submissions already posted:',
   },
   similarSubmissionsDescription: {
-    id: 'app.containers.ideas.similarIdeas.similarSubmissionsDescription',
+    id: 'app.components.ideas.similarIdeas.similarSubmissionsDescription',
     defaultMessage:
       'We found similar submisisons - engaging with them can help make them stronger!',
   },
   similarSubmissionsSearch: {
-    id: 'app.containers.ideas.similarIdeas.similarSubmissionsSearch',
+    id: 'app.components.ideas.similarIdeas.similarSubmissionsSearch',
     defaultMessage: 'Looking for similar submissions ...',
   },
   noSimilarSubmissions: {
-    id: 'app.containers.ideas.similarIdeas.noSimilarSubmissions',
+    id: 'app.components.ideas.similarIdeas.noSimilarSubmissions',
     defaultMessage: 'No similar submissions found.',
   },
 });
